yamcs-client/requirements.txt
protobuf
requests
websocket-client
